move Rich Text translation into utils/misc
