+++ src/@next/components/atoms/RichTextEditorContent/RichTextEditorContent.tsx
@@ -6,7 +6,7 @@ import EditorJS, {
 
 import { useIntl } from "react-intl";
 import { commonMessages } from "@temp/intl";
-import { hybritAbout, hybritProductDescriptions } from "@utils/misc";
+import { translateHybritText } from "@utils/misc";
 
 import Header from "@editorjs/header";
 import List from "@editorjs/list";
@@ -57,35 +57,7 @@ export const RichTextEditorContent: React.FC<RichTextEditorContentProps> = ({
     const formatBlock = (value: string) => {
       return { ...block, data: { text: value } };
     };
-    return text?.includes(hybritProductDescriptions.FS_BACKEND_DEV)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritFsBackEndDev))
-      : text?.includes(hybritProductDescriptions.FS_FRONTEND_DEV)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritFsFrontEndDev))
-      : text?.includes(hybritProductDescriptions.INTEG_DEV)
-      ? formatBlock(intl.formatMessage(commonMessages.integrationDev))
-      : text?.includes(hybritProductDescriptions.HYBRIT_BEER)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritBeerDescription))
-      : text?.includes(hybritProductDescriptions.HYBRIT_GINGER_BEER)
-      ? formatBlock(
-          intl.formatMessage(commonMessages.hybritGingerBeerDescription)
-        )
-      : text?.includes(hybritProductDescriptions.HYBRIT_PEN)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritPenDescription))
-      : text?.includes(hybritProductDescriptions.HYBRIT_SSD)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritExtSSDDescription))
-      : text?.includes(hybritProductDescriptions.HYBRIT_CUP)
-      ? formatBlock(
-          intl.formatMessage(commonMessages.hybritCoffeeCupDescription)
-        )
-      : text?.includes(hybritAbout.HYBRIT_ABOUT_HEADER)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritAboutHeader))
-      : text?.includes(hybritAbout.HYBRIT_ABOUT_P1)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritAboutP1))
-      : text?.includes(hybritAbout.HYBRIT_ABOUT_P2)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritAboutP2))
-      : text?.includes(hybritAbout.HYBRIT_ABOUT_P3)
-      ? formatBlock(intl.formatMessage(commonMessages.hybritAboutP3))
-      : { ...block, data: { text } };
+    return translateHybritText(text, intl, commonMessages, formatBlock, block);
   });
 
   const data = {
